refactor: Move UpdateRuleActiveRequestDTO to request-dto directory
- Move UpdateRuleActiveRequestDTO.ts to dto/request-dto/ subdirectory
- Update imports in RewriteRuleMessagingService.ts and IRewriteRuleMessagingPort.ts
- Update ESLint rule to enforce RequestDTO$ suffix for request-dto/ directory
- Update dto.md to document directory structure for request vs entity DTOs
