+++ docs/coding-standards/src/frameworks-and-drivers/messaging/dto.md
@@ -18,10 +18,10 @@
 
 ## 命名規約
 
-| 種別 | 命名パターン |
-|------|-------------|
-| リクエストDTO | `{操作名}RequestDTO` |
-| エンティティDTO | `{Entity}DTO` |
+| 種別 | 命名パターン | 配置ディレクトリ |
+|------|-------------|-----------------|
+| リクエストDTO | `{操作名}RequestDTO` | `dto/request-dto/` |
+| エンティティDTO | `{Entity}DTO` | `dto/` |
 
 > **参照**: [メッセージング層 コーディング規約](../messaging.md)
