+++ host-frontend-root/frontend-src-root/eslint-rules/clean-architecture/main.js
@@ -13,7 +13,7 @@ export default [
   ...applicationBusinessRulesDto,
   contentOnMessageReceivedUsecases,
   domain,
-  frameworksAndDriversMessagingDto,
+  ...frameworksAndDriversMessagingDto,
   frameworksAndDriversMessagingService,
   backgroundHandlers,
   contentHandlers,
