+++ host-frontend-root/frontend-src-root/src/frameworks-and-drivers/messaging/RewriteRuleMessagingService.ts
@@ -1,6 +1,6 @@
 import { GetByIdRequestDTO } from 'src/frameworks-and-drivers/messaging/dto/GetByIdRequestDTO';
 import { RewriteRuleDTO } from 'src/frameworks-and-drivers/messaging/dto/RewriteRuleDTO';
-import { UpdateRuleActiveRequestDTO } from 'src/frameworks-and-drivers/messaging/dto/UpdateRuleActiveRequestDTO';
+import { UpdateRuleActiveRequestDTO } from 'src/frameworks-and-drivers/messaging/dto/request-dto/UpdateRuleActiveRequestDTO';
 import { IRewriteRuleMessagingPort } from 'src/interface-adapters/ports/IRewriteRuleMessagingPort';
 
 /**
